Use Maven to build; restructure files; add README
