--- src/org/onebusaway/alexa/storage/ObaDynamoDbClient.java
@@ -1,48 +1,0 @@
-package org.onebusaway.alexa.storage;
-
-import com.amazonaws.services.dynamodbv2.AmazonDynamoDBClient;
-import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDBMapper;
-
-/**
- * Client for DynamoDB persistance layer for the OneBusAway skill.
- */
-public class ObaDynamoDbClient {
-	private final AmazonDynamoDBClient dynamoDBClient;
-
-	public ObaDynamoDbClient(final AmazonDynamoDBClient dynamoDBClient) {
-		this.dynamoDBClient = dynamoDBClient;
-	}
-
-	/**
-	 * Loads an item from DynamoDB by primary Hash Key. Callers of this method
-	 * should pass in an object which represents an item in the DynamoDB table
-	 * item with the primary key populated.
-	 * 
-	 * @param tableItem
-	 * @return
-	 */
-	public ObaUserDataItem loadItem(final ObaUserDataItem tableItem) {
-		DynamoDBMapper mapper = createDynamoDBMapper();
-		ObaUserDataItem item = mapper.load(tableItem);
-		return item;
-	}
-
-	/**
-	 * Stores an item to DynamoDB.
-	 * 
-	 * @param tableItem
-	 */
-	public void saveItem(final ObaUserDataItem tableItem) {
-		DynamoDBMapper mapper = createDynamoDBMapper();
-		mapper.save(tableItem);
-	}
-
-	/**
-	 * Creates a {@link DynamoDBMapper} using the default configurations.
-	 * 
-	 * @return
-	 */
-	private DynamoDBMapper createDynamoDBMapper() {
-		return new DynamoDBMapper(dynamoDBClient);
-	}
-}
