--- tst/org/onebusaway/alexa/LambdaFunctionHandlerTest.java
@@ -1,48 +1,0 @@
-package org.onebusaway.alexa;
-
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.OutputStream;
-
-import org.junit.BeforeClass;
-import org.junit.Test;
-
-import com.amazonaws.services.lambda.runtime.Context;
-
-/**
- * A simple test harness for locally invoking your Lambda function handler.
- */
-public class LambdaFunctionHandlerTest {
-
-    private static InputStream input;
-    private static OutputStream output;
-
-    @BeforeClass
-    public static void createInput() throws IOException {
-        // TODO: set up your sample input object here.
-        input = null;
-        output = null;
-    }
-
-    private Context createContext() {
-        TestContext ctx = new TestContext();
-
-        // TODO: customize your context here if needed.
-        ctx.setFunctionName("Your Function Name");
-
-        return ctx;
-    }
-
-    @Test
-    public void testLambdaFunctionHandler() {
-        LambdaFunctionHandler handler = new LambdaFunctionHandler();
-        Context ctx = createContext();
-
-//        Object output = handler.handleRequest(input, output, ctx);
-
-        // TODO: validate output here if needed.
-        if (output != null) {
-            System.out.println(output.toString());
-        }
-    }
-}
